Add /settings to index.
